backend/pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "REST API Workshop"
readme = "README.md"
license = { text = "Not open source" }
authors = [
  { name = "Jonathan Chung", email = "jonathan-chung@example.com" },
]
requires-python = "==3.13.*"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "argon2-cffi==25.1.0",
  "crispy-bootstrap5==2026.3",
  "django==6.0.3",
  "django-allauth[mfa]==65.15.0",
  "django-anymail[mailgun]==14.0",
  "django-cors-headers==4.9.0",
  "django-crispy-forms==2.6",
  "django-environ==0.13.0",
  "django-model-utils==5.0.0",
  "django-redis==6.0.0",
  "djangorestframework==3.17.0",
  "drf-spectacular==0.29.0",
  "gunicorn==25.1.0",
  "hiredis==3.3.1",
  "pillow==12.1.1",
  "python-slugify==8.0.4",
  "redis==7.4.0",
  "whitenoise==6.12.0",
]

[dependency-groups]
dev = [
  "coverage==7.13.5",
  "django-coverage-plugin==3.2.0",
  "django-debug-toolbar==6.2.0",
  "django-extensions==4.1",
  "django-stubs[compatible-mypy]==6.0.1",
  "djangorestframework-stubs==3.16.8",
  "djlint==1.36.4",
  "factory-boy==3.3.3",
  "ipdb==0.13.13",
  "mypy==1.19.1",
  "pre-commit==4.5.1",
  "psycopg[binary]==3.3.3",
  "pytest==9.0.2",
  "pytest-django==4.12.0",
  "pytest-sugar==1.1.1",
  "ruff==0.15.7",
  "sphinx==9.1.0",
  "sphinx-autobuild==2025.8.25",
  "werkzeug[watchdog]==3.1.7",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
  "*/migrations/*.py",
  "staticfiles/*",
]

lint.select = [
  "A",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "DJ",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "EXE",
  "F",
  "FA",
  "FBT",
  "FLY",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "N",
  "PD",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "SIM102", # sometimes it's better to nest
  # of types for comparison.
  # Deactivated because it can make the code slow:
  # https://github.com/astral-sh/ruff/issues/7871
]
lint.isort.force-single-line = true

[tool.pyproject-fmt]
keep_full_version = true

# ==== pytest ====

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
  "tests.py",
  "test_*.py",
]

# ==== Coverage ====

[tool.coverage.run]
include = [ "core/**" ]
omit = [ "*/migrations/*", "*/tests/*" ]
plugins = [ "django_coverage_plugin" ]

# ==== mypy ====

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====

[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2
